Application/requirements.txt
attrdict
nvidia-pyindex
tritonclient[all]
opencv-python
matplotlib
numpy
pillow